more warnings
svn:sync-xref-src-repo-rev: r37321

+++ src/VBox/Devices/vl_vbox.h
@@ -1,4 +1,4 @@
-/* $Id: vl_vbox.h 12640 2008-09-22 13:19:30Z knut.osmundsen@oracle.com $ */
+/* $Id: vl_vbox.h 12910 2008-10-02 09:36:48Z noreply@oracle.com $ */
 /** @file
  * VBox vl.h Replacement.
  *
@@ -145,14 +145,14 @@ DECLINLINE(unsigned) qemu_get_be32(QEMUFile *f)
 {
     uint32_t u32;
     int rc = SSMR3GetU32(f, &u32);
-    return RT_SUCCESS(rc) ? u32 : ~0;
+    return RT_SUCCESS(rc) ? u32 : ~0U;
 }
 
 DECLINLINE(int64_t) qemu_get_be64(QEMUFile *f)
 {
     uint64_t u64;
     int rc = SSMR3GetU64(f, &u64);
-    return RT_SUCCESS(rc) ? u64 : ~0;
+    return RT_SUCCESS(rc) ? (int64_t)u64 : ~0;
 }
 
 #define qemu_put_timer(f, ts)   TMR3TimerSave((ts), (f))
